+++ Client/GameProc/DelayedExp.h
@@ -6,8 +6,8 @@
 
 struct EXP_DATA {
   int iOwnerIndex;
-  int iAddExp;
-  int iExp;
+  int64_t iAddExp;
+  int64_t iExp;
   int iGetTime; /// 서버로 부터 받은시간. 
 
   int iStamina; /// 스태미너도 현재 스태미너가 서버에서 날라오기때문에..
@@ -28,7 +28,7 @@ public:
   CDelayedExp(void );
   ~CDelayedExp(void);
 
-  void PushEXPData(int iOwnerIndex, int iAddExp, int iEXP, int iStamina);
+  void PushEXPData(int iOwnerIndex, int64_t iAddExp, int64_t iEXP, int iStamina);
   void GetExp(int      iOwnerIndex);
 
   void Proc();
